requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
streamlit>=1.28.0

# HTTP Client
httpx>=0.24.0
requests>=2.31.0

# Data Validation
pydantic>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# Async Support
asyncio>=3.4.3
